Introduce maps/coord.py
